椭圆的构造,显示,边线检测,内部检测等均已完成.目前和其他组件一样不支持旋转和缩放等

+++ lib/model/geometry/lines/ray.dart
@@ -11,7 +11,7 @@ class Ray{
   Vector2D getVector(){
     return Vector2D(anyPointOnLine.x - start.x, anyPointOnLine.y - start.y);
   }
-  double getAngle(){
+  Decimal getAngle(){
     return getVector().getAngle();
   }
   void normalize(){
